Remove mocks from repository mocker Add separated Mocks for new entities
