+++ Streetcode/Streetcode.XUnitTest/MediatRTests/Mocks/RepositoryMocker.cs
@@ -871,205 +871,5 @@ namespace Streetcode.XUnitTest.MediatRTests.Mocks
 
             return mockRepo;
         }
-
-        public static Mock<IRepositoryWrapper> GetDictionaryItemRepositoryMock()
-        {
-            var mockRepo = new Mock<IRepositoryWrapper>();
-
-            var dictionaryItems = new List<DictionaryItem>()
-            {
-                new DictionaryItem { Id = 1, Name = "First Dictionary", Description = "First Description"},
-                new DictionaryItem { Id = 2, Name = "Second Dictionary", Description = "Second Description"},
-                new DictionaryItem { Id = 3, Name = "Third Dictionary", Description = "Third Description"},
-                new DictionaryItem { Id = 4, Name = "Fourth Dictionary", Description = "Fourth Description"},
-            };
-
-            mockRepo.Setup(x => x.DictionaryItemRepository.GetAllAsync(It.IsAny<Expression<Func<DictionaryItem, bool>>>(), It.IsAny<Func<IQueryable<DictionaryItem>, IIncludableQueryable<DictionaryItem, object>>>()))
-                .ReturnsAsync(dictionaryItems);
-
-            mockRepo.Setup(x => x.DictionaryItemRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<DictionaryItem, bool>>>(), It.IsAny<Func<IQueryable<DictionaryItem>, IIncludableQueryable<DictionaryItem, object>>>()))
-                .ReturnsAsync((Expression<Func<DictionaryItem, bool>> predicate, Func<IQueryable<DictionaryItem>, IIncludableQueryable<DictionaryItem, object>> include) =>
-                {
-                    return dictionaryItems.FirstOrDefault(predicate.Compile());
-                });
-
-            return mockRepo;
-        }
-
-        public static Mock<IRepositoryWrapper> GetArticleRepositoryMock()
-        {
-            var mockRepo = new Mock<IRepositoryWrapper>();
-
-            var articles = new List<Article>()
-            {
-                new Article { Id = 1, Title = "First Title", Text = "First Text"},
-                new Article { Id = 2, Title = "Second Title", Text = "Second Text"},
-                new Article { Id = 3, Title = "Third Title", Text = "Third Text"},
-                new Article { Id = 4, Title = "Fourth Title", Text = "Fourth Text"},
-            };
-
-            mockRepo.Setup(x => x.ArticleRepository.GetAllAsync(It.IsAny<Expression<Func<Article, bool>>>(), It.IsAny<Func<IQueryable<Article>, IIncludableQueryable<Article, object>>>()))
-                .ReturnsAsync(articles);
-
-            mockRepo.Setup(x => x.ArticleRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<Article, bool>>>(), It.IsAny<Func<IQueryable<Article>, IIncludableQueryable<Article, object>>>()))
-                .ReturnsAsync((Expression<Func<Article, bool>> predicate, Func<IQueryable<Article>, IIncludableQueryable<Article, object>> include) =>
-                {
-                    return articles.FirstOrDefault(predicate.Compile());
-                });
-
-            return mockRepo;
-        }
-
-        public static Mock<IRepositoryWrapper> GetInfoBlockRepositoryMock()
-        {
-            var mockRepo = new Mock<IRepositoryWrapper>();
-
-            var infoBlocks = new List<InfoBlock>()
-            {
-                new InfoBlock
-                {
-                    Id = 1,
-                    ArticleId = 1,
-                    Article = new Article { Id = 1, Text = "First Text", Title = "First Title" },
-                    VideoURL = "www.youtube.com",
-                    AuthorShipId = 1,
-                    AuthorShip = new AuthorShip
-                    {
-                        Id = 1,
-                        Text = "First Text",
-                        AuthorShipHyperLinkId = 1,
-                        AuthorShipHyperLink = new AuthorShipHyperLink { Id = 1, Title = "First Title", URL = "First URL"},
-                    },
-                },
-                new InfoBlock
-                {
-                    Id = 2,
-                    ArticleId = 2,
-                    Article = new Article { Id = 2, Text = "Second Text", Title = "Second Title" },
-                    VideoURL = "www.youtube2.com",
-                    AuthorShipId = 2,
-                    AuthorShip = new AuthorShip
-                    {
-                        Id = 2,
-                        Text = "Second Text",
-                        AuthorShipHyperLinkId = 2,
-                        AuthorShipHyperLink = new AuthorShipHyperLink { Id = 2, Title = "Second Title", URL = "Second URL"},
-                    },
-                },
-                new InfoBlock
-                {
-                    Id = 3,
-                    ArticleId = 3,
-                    Article = new Article { Id = 3, Text = "Third Text", Title = "Third Title" },
-                    VideoURL = "www.youtube3.com",
-                    AuthorShipId = 3,
-                    AuthorShip = new AuthorShip
-                    {
-                        Id = 3,
-                        Text = "Third Text",
-                        AuthorShipHyperLinkId = 3,
-                        AuthorShipHyperLink = new AuthorShipHyperLink { Id = 3, Title = "Third Title", URL = "Third URL"},
-                    },
-                },
-                new InfoBlock
-                {
-                    Id = 4,
-                    ArticleId = 4,
-                    Article = new Article { Id = 4, Text = "Fourth Text", Title = "Fourth Title" },
-                    VideoURL = "www.youtube4.com",
-                    AuthorShipId = 4,
-                    AuthorShip = new AuthorShip
-                    {
-                        Id = 4,
-                        Text = "Fourth Text",
-                        AuthorShipHyperLinkId = 4,
-                        AuthorShipHyperLink = new AuthorShipHyperLink { Id = 4, Title = "Fourth Title", URL = "Fourth URL"},
-                    },
-                },
-            };
-
-            mockRepo.Setup(x => x.InfoBlockRepository.GetAllAsync(It.IsAny<Expression<Func<InfoBlock, bool>>>(), It.IsAny<Func<IQueryable<InfoBlock>, IIncludableQueryable<InfoBlock, object>>>()))
-                .ReturnsAsync(infoBlocks);
-
-            mockRepo.Setup(x => x.InfoBlockRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<InfoBlock, bool>>>(), It.IsAny<Func<IQueryable<InfoBlock>, IIncludableQueryable<InfoBlock, object>>>()))
-                .ReturnsAsync((Expression<Func<InfoBlock, bool>> predicate, Func<IQueryable<InfoBlock>, IIncludableQueryable<InfoBlock, object>> include) =>
-                {
-                    return infoBlocks.FirstOrDefault(predicate.Compile());
-                });
-
-            return mockRepo;
-        }
-
-        public static Mock<IRepositoryWrapper> GetAuthorShipRepositoryMock()
-        {
-            var mockRepo = new Mock<IRepositoryWrapper>();
-
-            var authorShips = new List<AuthorShip>()
-            {
-                new AuthorShip
-                {
-                    Id = 1,
-                    Text = "First Text",
-                    AuthorShipHyperLinkId = 1,
-                    AuthorShipHyperLink = new AuthorShipHyperLink { Id = 1, Title = "First Title", URL = "First URL"},
-                },
-                new AuthorShip
-                {
-                    Id = 2,
-                    Text = "Second Text",
-                    AuthorShipHyperLinkId = 2,
-                    AuthorShipHyperLink = new AuthorShipHyperLink { Id = 2, Title = "Second Title", URL = "Second URL"},
-                },
-                new AuthorShip
-                {
-                    Id = 3,
-                    Text = "Third Text",
-                    AuthorShipHyperLinkId = 3,
-                    AuthorShipHyperLink = new AuthorShipHyperLink { Id = 3, Title = "Third Title", URL = "Third URL"},
-                },
-                new AuthorShip
-                {
-                    Id = 4,
-                    Text = "Fourth Text",
-                    AuthorShipHyperLinkId = 4,
-                    AuthorShipHyperLink = new AuthorShipHyperLink { Id = 4, Title = "Fourth Title", URL = "Fourth URL"},
-                },
-            };
-
-            mockRepo.Setup(x => x.AuthorShipRepository.GetAllAsync(It.IsAny<Expression<Func<AuthorShip, bool>>>(), It.IsAny<Func<IQueryable<AuthorShip>, IIncludableQueryable<AuthorShip, object>>>()))
-                .ReturnsAsync(authorShips);
-
-            mockRepo.Setup(x => x.AuthorShipRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<AuthorShip, bool>>>(), It.IsAny<Func<IQueryable<AuthorShip>, IIncludableQueryable<AuthorShip, object>>>()))
-                .ReturnsAsync((Expression<Func<AuthorShip, bool>> predicate, Func<IQueryable<AuthorShip>, IIncludableQueryable<AuthorShip, object>> include) =>
-                {
-                    return authorShips.FirstOrDefault(predicate.Compile());
-                });
-
-            return mockRepo;
-        }
-
-        public static Mock<IRepositoryWrapper> GetAuthorShipHyperLinkRepositoryMock()
-        {
-            var mockRepo = new Mock<IRepositoryWrapper>();
-
-            var authorShipHyperLinks = new List<AuthorShipHyperLink>()
-            {
-                new AuthorShipHyperLink { Id = 1, Title = "First Title", URL = "First URL"},
-                new AuthorShipHyperLink { Id = 2, Title = "Second Title", URL = "Second URL"},
-                new AuthorShipHyperLink { Id = 3, Title = "Third Title", URL = "Third URL"},
-                new AuthorShipHyperLink { Id = 4, Title = "Fourth Title", URL = "Fourth URL"},
-            };
-
-            mockRepo.Setup(x => x.AuthorShipHyperLinkRepository.GetAllAsync(It.IsAny<Expression<Func<AuthorShipHyperLink, bool>>>(), It.IsAny<Func<IQueryable<AuthorShipHyperLink>, IIncludableQueryable<AuthorShipHyperLink, object>>>()))
-                .ReturnsAsync(authorShipHyperLinks);
-
-            mockRepo.Setup(x => x.AuthorShipHyperLinkRepository.GetFirstOrDefaultAsync(It.IsAny<Expression<Func<AuthorShipHyperLink, bool>>>(), It.IsAny<Func<IQueryable<AuthorShipHyperLink>, IIncludableQueryable<AuthorShipHyperLink, object>>>()))
-                .ReturnsAsync((Expression<Func<AuthorShipHyperLink, bool>> predicate, Func<IQueryable<AuthorShipHyperLink>, IIncludableQueryable<AuthorShipHyperLink, object>> include) =>
-                {
-                    return authorShipHyperLinks.FirstOrDefault(predicate.Compile());
-                });
-
-            return mockRepo;
-        }
     }
 }
